[#193] feat : 프로필 페이지 서버렌더링
